Refactor clock raw counter backends
