+++ core/cu29_clock/src/raw_counter/aarch64_cntvct.rs
@@ -1,3 +1,6 @@
+// Source: ARM generic timer virtual count register (`CNTVCT_EL0`).
+// Guarantee: monotonic fixed-frequency counter when firmware/OS enables EL0
+// access. The counter is system-level rather than a per-thread wall clock.
 pub fn initialize() {}
 
 #[inline(always)]
