+++ core/cu29_clock/src/raw_counter/wasm_instant.rs
@@ -1,5 +1,8 @@
 use std::sync::OnceLock;
 
+// Source: `web_time::Instant` elapsed from a process-local start instant.
+// Guarantee: monotonic elapsed time within the current WASM instance. Values are
+// already nanoseconds, so calibration should converge to a 1:1 rate.
 static START: OnceLock<web_time::Instant> = OnceLock::new();
 
 #[inline(always)]
